Change Cache system

+++ frontend/controllers/SiteController.php
@@ -289,7 +289,7 @@ class SiteController extends Controller
 
         if($prod_attr_query == '' && $searchword == ''){
             $x =  PartnersProductsToCategories::find()->select('MAX(products.`products_last_modified`) as products_last_modifieds ')->distinct()->JoinWith('products')->where('categories_id IN ('.$cat.')')->asArray()->one();
-            $prod =  PartnersProductsToCategories::find()->select('products.products_id as prod,  products.products_last_modified as last ')->JoinWith('products')->where('  categories_id IN ('.$cat.') and products_status = 1    and products.products_quantity > 0    and products.removable != 1    and products_price <= :end_price and products_price >= :start_price  and products.manufacturers_id NOT IN ('.$hide_man.')',[':start_price' => $start_price, ':end_price' => $end_price])->limit($count)->offset($start_arr)->JoinWith('productsAttributesDescr')->JoinWith('productsDescription')->distinct()->orderBy($orders)->asArray()->all();
+            $prod =  PartnersProductsToCategories::find()->select('products.products_id as prod,  products.products_last_modified as last ')->JoinWith('products')->where('  categories_id IN ('.$cat.') and products_status = 1    and products.products_quantity > 0    and products.removable != 1    and products_price <= :end_price and products_price >= :start_price  and products.manufacturers_id NOT IN ('.$hide_man.')',[':start_price' => $start_price, ':end_price' => $end_price])->limit($count)->offset($start_arr)->JoinWith('productsDescription')->distinct()->orderBy($orders)->asArray()->all();
             foreach($prod as $values){
                 $dataprod = Yii::$app->cache->get(urlencode('prod-'.$values['prod']));
                 if(isset($dataprod) && $dataprod['last'] == $values['last']){
@@ -302,10 +302,12 @@ class SiteController extends Controller
                 $prodarr = implode(',', $nodata);
                 $datar = PartnersProductsToCategories::find()->JoinWith('products')->where('  categories_id IN ('.$cat.') and products_status = 1    and products.products_quantity > 0    and products.removable != 1    and products_price <= :end_price and products_price >= :start_price  and products.manufacturers_id NOT IN ('.$hide_man.') and products.products_id IN ('.$prodarr.')',[':start_price' => $start_price, ':end_price' => $end_price])->JoinWith('productsDescription')->JoinWith('productsAttributes')->JoinWith('productsAttributesDescr')->distinct()->asArray()->all();
                 foreach($datar as $valuesr){
-                    Yii::$app->cache->delete(urlencode('prod-'.$valuesr['products_id']));
-                    Yii::$app->cache->set(urlencode('prod-'.$valuesr['products_id']), ['data' => $valuesr, 'last' => $valuesr['products']['products_last_modified']]);
-                    $data[] = $valuesr;
-                }
+                  if(isset($valuesr['productsDescription'])) {
+                      Yii::$app->cache->delete(urlencode('prod-' . $valuesr['products_id']));
+                      Yii::$app->cache->set(urlencode('prod-' . $valuesr['products_id']), ['data' => $valuesr, 'last' => $valuesr['products']['products_last_modified']]);
+                      $data[] = $valuesr;
+                  }
+                  }
             }
 
             if(!isset($x['products_last_modifieds'])){
@@ -319,7 +321,7 @@ class SiteController extends Controller
                 Yii::$app->cache->delete(urlencode('first-'.$cat_start.'-'.$hide_man.'-'.$start_price.'-'.$end_price.'-'.$count.'-page'.$page.'-'.$sort));
             }
             if ($dataque === false || ($checkcache !== $dataque['checkcache'])) {
-                $productattrib = PartnersProductsToCategories::find()->select(['products_options_values.products_options_values_id','products_options_values.products_options_values_name'])->distinct()->JoinWith('products')->where('categories_id IN ('.$cat.')  and products.removable != 1   and products.products_quantity > 0   and products_status=1  and products_price <= :end_price and products_price >= :start_price  and products.manufacturers_id NOT IN ('.$hide_man.')  ',[':start_price' => $start_price, ':end_price' => $end_price])->orderBy('`products_price` DESC')->JoinWith('productsDescription')->JoinWith('productsAttributes')->JoinWith('productsAttributesDescr')->asArray()->all();
+                $productattrib = PartnersProductsToCategories::find()->select(['products_options_values.products_options_values_id','products_options_values.products_options_values_name'])->distinct()->JoinWith('products')->where('categories_id IN ('.$cat.')  and products.removable != 1   and products.products_quantity > 0   and products_status=1  and products_price <= :end_price and products_price >= :start_price  and products.manufacturers_id NOT IN ('.$hide_man.')  ',[':start_price' => $start_price, ':end_price' => $end_price])->JoinWith('productsDescription')->JoinWith('productsAttributes')->JoinWith('productsAttributesDescr')->asArray()->all();
                 $count_arrs = PartnersProductsToCategories::find()->JoinWith('products')->where('categories_id IN ('.$cat.') and products_status=1  and products.products_quantity > 0   and products_price <= :end_price and products.removable != 1   and products_price >= :start_price  and products.manufacturers_id NOT IN ('.$hide_man.')',[':start_price' => $start_price, ':end_price' => $end_price])->distinct()->count();
                 $price_max = PartnersProductsToCategories::find()->select('MAX(`products_price`) as maxprice')->distinct()->JoinWith('products')->where('categories_id IN ('.$cat.')     and products.products_quantity > 0    and products.removable != 1    and products_status=1 and products.manufacturers_id NOT IN ('.$hide_man.') ')->asArray()->one();
                 Yii::$app->cache->set(urlencode('first-'.$cat_start.'-'.$hide_man.'-'.$start_price.'-'.$end_price.'-'.$count.'-page'.$page.'-'.$sort), ['productattrib' => $productattrib, 'count_arrs' => $count_arrs, 'price_max' =>  $price_max, 'checkcache' => $checkcache]);
